% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_get_backup_vault_notifications}
\alias{backup_get_backup_vault_notifications}
\title{Returns event notifications for the specified backup vault}
\usage{
backup_get_backup_vault_notifications(BackupVaultName)
}
\arguments{
\item{BackupVaultName}{[required] The name of a logical container where backups are stored. Backup vaults
are identified by names that are unique to the account used to create
them and the AWS Region where they are created. They consist of
lowercase letters, numbers, and hyphens.}
}
\description{
Returns event notifications for the specified backup vault.
}
\section{Request syntax}{
\preformatted{backup$get_backup_vault_notifications(
  BackupVaultName = "string"
)
}
}

\keyword{internal}
